[test-suite] Rename firebase tests and make them less flaky
