+++ apps/test-suite/TestModules.js
@@ -70,8 +70,8 @@ export function getTestModules() {
     require('./tests/FirebaseCore'),
     require('./tests/FirebaseAnalytics'),
     require('./tests/FirebaseRecaptcha'),
+    require('./tests/FirebaseJSSDKCompat'),
     require('./tests/FirebaseJSSDK'),
-    require('./tests/FirebaseJSSDKv9'),
     require('./tests/ImageManipulator'),
     optionalRequire(() => require('./tests/SQLite'))
   );
